auto-claude: subtask-6-1 - 简化主屏幕文件 - 移除已提取的类定义
- 将 VibeLibraryScreen 从 1432 行简化到 318 行
- 创建 DesktopVibeLibraryLayout 桌面布局组件
- 创建 MobileVibeLibraryLayout 移动端布局组件
- 创建 VibePaginationBar 分页条组件
- 创建 VibeDropOverlay 拖拽覆盖层组件
- 创建 VibeLibraryActionsMixin 集中处理操作逻辑
- 遵循 generation_screen.dart 的简化模式
