+++ lib/presentation/screens/vibe_library/widgets/vibe_pagination_bar.dart
@@ -1,0 +1,84 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import '../../../providers/vibe_library_provider.dart';
+
+/// Vibe库分页条组件
+class VibePaginationBar extends ConsumerWidget {
+  const VibePaginationBar({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final state = ref.watch(vibeLibraryNotifierProvider);
+    final theme = Theme.of(context);
+
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+      decoration: BoxDecoration(
+        color: theme.colorScheme.surfaceContainerLow,
+        border: Border(
+          top: BorderSide(
+            color: theme.colorScheme.outlineVariant.withOpacity(0.3),
+          ),
+        ),
+      ),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          IconButton(
+            icon: const Icon(Icons.chevron_left),
+            onPressed: state.currentPage > 0
+                ? () {
+                    ref
+                        .read(vibeLibraryNotifierProvider.notifier)
+                        .loadPreviousPage();
+                  }
+                : null,
+          ),
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 16),
+            child: Text(
+              '${state.currentPage + 1} / ${state.totalPages} 页',
+              style: theme.textTheme.bodyMedium,
+            ),
+          ),
+          IconButton(
+            icon: const Icon(Icons.chevron_right),
+            onPressed: state.currentPage < state.totalPages - 1
+                ? () {
+                    ref
+                        .read(vibeLibraryNotifierProvider.notifier)
+                        .loadNextPage();
+                  }
+                : null,
+          ),
+          const SizedBox(width: 16),
+          Text('每页:', style: theme.textTheme.bodySmall),
+          const SizedBox(width: 8),
+          DropdownButton<int>(
+            value: state.pageSize,
+            underline: const SizedBox(),
+            items: [20, 50, 100].map((size) {
+              return DropdownMenuItem(
+                value: size,
+                child: Text('$size'),
+              );
+            }).toList(),
+            onChanged: (value) {
+              if (value != null) {
+                ref
+                    .read(vibeLibraryNotifierProvider.notifier)
+                    .setPageSize(value);
+              }
+            },
+          ),
+          const Spacer(),
+          Text(
+            '共 ${state.filteredCount} 个Vibe',
+            style: theme.textTheme.bodySmall,
+          ),
+        ],
+      ),
+    );
+  }
+}
